- fingerprint metrics WIP
